move ReifiedEvent to Common, reverse Dependency to EventDriver, to put upper bound on Result's Type parameter.
