+++ bundles/org.palladiosimulator.analyzer.slingshot.common/META-INF/MANIFEST.MF
@@ -10,7 +10,6 @@ Require-Bundle: org.palladiosimulator.commons;bundle-version="5.1.0",
  org.eclipse.core.resources;bundle-version="3.16.0",
  org.eclipse.jdt.annotation;bundle-version="2.2.600",
  com.google.guava;bundle-version="30.1.0",
- org.eclipse.emf.ecore,
- org.palladiosimulator.analyzer.slingshot.eventdriver
+ org.eclipse.emf.ecore
 Export-Package: org.palladiosimulator.analyzer.slingshot.common.events
 Import-Package: org.eclipse.emf.ecore.util
