+++ bundles/org.palladiosimulator.analyzer.slingshot.eventdriver/META-INF/MANIFEST.MF
@@ -9,7 +9,8 @@ Bundle-ClassPath: lib/rxjava-3.1.5.jar,
  .
 Require-Bundle: org.reactivestreams;bundle-version="1.0.2",
  org.apache.log4j,
- com.google.guava
+ com.google.guava,
+ org.palladiosimulator.analyzer.slingshot.common
 Export-Package: org.palladiosimulator.analyzer.slingshot.eventdriver,
  org.palladiosimulator.analyzer.slingshot.eventdriver.annotations,
  org.palladiosimulator.analyzer.slingshot.eventdriver.annotations.eventcontract,
