+++ bundles/org.palladiosimulator.analyzer.slingshot.eventdriver/src/org/palladiosimulator/analyzer/slingshot/eventdriver/returntypes/Result.java
@@ -7,6 +7,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
+import org.palladiosimulator.analyzer.slingshot.common.events.SlingshotEvent;
+
 /**
  * A result is a container that holds any type of information of an event handler's result.
  * This could be, for example, new events that should be scheduled afterwards, especially
@@ -18,7 +20,7 @@ import java.util.Set;
  *
  * @param <T> The containers type
  */
-public final class Result<T> {
+public final class Result<T extends SlingshotEvent> {
 
 	private Set<T> resultEvents;
 
@@ -36,15 +38,15 @@ public final class Result<T> {
 		return Collections.unmodifiableSet(this.resultEvents);
 	}
 
-	public static <T> Result<T> of(final T... events) {
+	public static <T extends SlingshotEvent> Result<T> of(final T... events) {
 		return new Result<>(List.of(events));
 	}
 
-	public static <T, S extends T> Result<T> of(final Collection<S> resultEvents) {
+	public static <T extends SlingshotEvent, S extends T> Result<T> of(final Collection<S> resultEvents) {
 		return new Result<>(List.copyOf(resultEvents));
 	}
 
-	public static <T, S extends T> Result<T> of(final Optional<S> resultEvent) {
+	public static <T extends SlingshotEvent, S extends T> Result<T> of(final Optional<S> resultEvent) {
 		if (resultEvent.isPresent()) {
 			return new Result<>(List.of(resultEvent.get()));
 		}
@@ -52,12 +54,12 @@ public final class Result<T> {
 	}
 
 	@Deprecated
-	public static <T> Result<T> from(final Collection<T> resultEvents) {
+	public static <T extends SlingshotEvent> Result<T> from(final Collection<T> resultEvents) {
 		return new Result<>(resultEvents);
 	}
 
 	@Deprecated
-	public static <T> Result<T> empty() {
+	public static <T extends SlingshotEvent> Result<T> empty() {
 		return new Result<>(List.of());
 	}
 }
